requirements.txt
Django>=3.1, <3.2
python-dotenv>=0.14.0, <0.15.0
djangorestframework>=3.11.1, <3.12.0
drf-writable-nested>=0.6.1,<0.7.0
